Syntastic used only in vim

+++ .vim/rc/ft-vim+vimrc.vim
@@ -2,33 +2,36 @@
 
 function! s:VimSupport()
     " syntax checker: set syntastic to use 'vint'                      {{{1
-    " - ensure it is installed
-    let l:check_cmd = 'python -c "import vint"'
-    call system(l:check_cmd)
-    if v:shell_error
-        let l:install_cmd = 'pip install --upgrade vim-vint'
-        let l:feedback = systemlist(l:install_cmd)
+    " - syntastic is used in vim
+    if exists(':shell')
+        " - ensure vint is installed
+        let l:check_cmd = 'python -c "import vint"'
+        call system(l:check_cmd)
         if v:shell_error
-            echoerr 'Unable to install vint binary'
-            if len(l:feedback)
-                echoerr 'Shell feedback:'
-                for l:line in l:feedback
-                    echoerr '  ' . l:line
-                endfor
+            let l:install_cmd = 'pip install --upgrade vim-vint'
+            let l:feedback = systemlist(l:install_cmd)
+            if v:shell_error
+                echoerr 'Unable to install vint binary'
+                if len(l:feedback)
+                    echoerr 'Shell feedback:'
+                    for l:line in l:feedback
+                        echoerr '  ' . l:line
+                    endfor
+                endif
             endif
         endif
-    endif
-    call system(l:check_cmd)
-    if !v:shell_error
-        if !exists('g:syntastic_vim_checkers')
-            let g:syntastic_vim_checkers = []
+        call system(l:check_cmd)
+        if !v:shell_error
+            if !exists('g:syntastic_vim_checkers')
+                let g:syntastic_vim_checkers = []
+            endif
+            if !count(g:syntastic_vim_checkers, 'vint')
+                call add(g:syntastic_vim_checkers, 'vint')
+            endif
+        else
+            echoerr "Vim syntax checker 'vint' is not available"
         endif
-        if !count(g:syntastic_vim_checkers, 'vint')
-            call add(g:syntastic_vim_checkers, 'vint')
-        endif
-    else
-        echoerr "Vim syntax checker 'vint' is not available"
-    endif                                                            " }}}1
+    endif                                                            " }}}1"
     " K command: use internal vim help                                 {{{1
     setlocal keywordprg=:help                                        " }}}1
 endfunction
